requirements.txt
# Lead Finder AI - Development Requirements
# ==========================================
# Para desarrollo local con SQLite

# Backend Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
flask-bcrypt>=1.0.1
flask-wtf>=1.2.0
flask-limiter>=3.5.0

# API & Validation
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Stripe Payments
stripe>=7.0.0

# Scheduler
apscheduler>=3.10.0

# Web Scraping
praw>=7.7.0
beautifulsoup4>=4.12.0
httpx>=0.25.0

# AI Integration
openai>=1.6.0

# Utils
python-dateutil>=2.8.0
pytz>=2023.3
gunicorn>=21.0.0
cryptography>=41.0.0

# Email Validation (PRO)
dnspython>=2.4.0

# Resilience & Error Tracking
tenacity>=9.0.0
sentry-sdk>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
